+++ tests/commands-anx/version-anx.test.ts
@@ -6,6 +6,7 @@ describe('versionAnx', () => {
     const myAnx = anx.createAccessor('');
     const rslt = await myAnx.versionAnx();
     expect(rslt.exitCode).toBe(0);
+    expect(rslt.out).toContain('build flags:');
     expect(rslt.out).toContain('git-annex version:');
     expect(rslt.out).toContain('key/value backends:');
   });
@@ -15,6 +16,7 @@ describe('versionAnx', () => {
     const rslt = await myAnx.versionAnx({ '--raw': null });
     expect(rslt.exitCode).toBe(0);
     expect(rslt.out.length).toBeGreaterThan(0);
+    expect(rslt.out).not.toContain('build flags:');
     expect(rslt.out).not.toContain('git-annex version:');
     expect(rslt.out).not.toContain('key/value backends:');
   });